+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Missing lemgrams in autocomplete [#416](https://github.com/spraakbanken/korp-frontend/issues/416)
 - The response JSON download button now handles POST and logged-in requests, and has been moved into each corresponding result tab [#417](https://github.com/spraakbanken/korp-frontend/issues/417)
 - Limit search history selector width [#415](https://github.com/spraakbanken/korp-frontend/issues/415)
+- Search not triggered when choosing simple lemgram search from search history [#152](https://github.com/spraakbanken/korp-frontend/issues/152)
 
 ## [9.7.2] - 2024-12-09